# Hook up net/trace with opentracing