--- a/cli/cli_test.go
+++ b/cli/cli_test.go
@@ -55,7 +55,7 @@ func newCLITest() cliTest {
 	// Reset the client context for each test. We don't reset the
 	// pointer (because they are tied into the flags), but instead
 	// overwrite the existing struct's values.
-	context.InitDefaults()
+	cliContext.InitDefaults()
 
 	osStderr = os.Stdout