--- a/cli/main_test.go
+++ b/cli/main_test.go
@@ -22,12 +22,10 @@ import (
 	"github.com/cockroachdb/cockroach/security"
 	"github.com/cockroachdb/cockroach/security/securitytest"
 	"github.com/cockroachdb/cockroach/util/leaktest"
-	"github.com/cockroachdb/cockroach/util/tracing"
 )
 
 func init() {
 	security.SetReadFileFn(securitytest.Asset)
-	tracing.SetTestTracing()
 }
 
 //go:generate ../util/leaktest/add-leaktest.sh *_test.go