--- a/client/main_test.go
+++ b/client/main_test.go
@@ -22,11 +22,9 @@ import (
 	"github.com/cockroachdb/cockroach/security"
 	"github.com/cockroachdb/cockroach/security/securitytest"
 	"github.com/cockroachdb/cockroach/util/leaktest"
-	"github.com/cockroachdb/cockroach/util/tracing"
 )
 
 func init() {
-	tracing.SetTestTracing()
 	security.SetReadFileFn(securitytest.Asset)
 }