There is still some upstream movement, so for now
a sensible option is to log directly to net/trace.

Removed disabling of tracers in benchmarks: We'll
run tracing in real world loads, so it's good to
have it on the radar and to make sure the performance
impact is palatable.

Apologies for the unrelated-looking renames.

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.svg" height="40" alt="Review on Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/4385)
<!-- Reviewable:end -->