@petermattis 2016-02-15
Are traces sampled now? If not, this is going to distort benchmarks and yet not be indicative of eventual performance. 

What is the performance impact of this change? In particular, I'm interested in the `(Insert|Update|Delete|Scan).*Cockroach` benchmarks in the `sql` package. Run with `go test -count 10` and then use `benchstat` to compare against master.

---

Review status: 0 of 31 files reviewed at latest revision, 3 unresolved discussions, some commit checks failed.

---

*[kv/dist_sender.go, line 228 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAaebSMTMUYNm6z1Gyh:-KAaebSNXaeYMfjoJV1B:1260808254) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/kv/dist_sender.go#L228)):*
You also need to pass a context in to the `RangeLookup` method so that the span can be associated with the correct trace, right?

---

*[storage/replica.go, line 786 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAaf6pbVFYvSUXUaEuh:-KAaf6pbVFYvSUXUaEui:543530695) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/storage/replica.go#L786)):*
Perhaps replace the `fmt.Sprintf` with `reflect.TypeOf(args).String()`. I think they are equivalent, though the latter avoids the `fmt` machinery.

---

*[util/tracing/nettrace.go, line 92 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAaflAhYXdkJC9Wehmu:-KAaflAhYXdkJC9Wehmv:-199689173) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/util/tracing/nettrace.go#L92)):*
`s/ntwi/ntwe/g`?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-:-KAag3hNS5zMqjcDryMf:-1867755225)*
<!-- Sent from Reviewable.io -->