@tamird 2016-02-15
Reviewed 31 of 31 files at r1.
Review status: all files reviewed at latest revision, 8 unresolved discussions, some commit checks failed.

---

*[kv/local_test_cluster.go, line 50 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAaheMsY5dExBzkpdyD:-KAaheMsY5dExBzkpdyE:2124961268) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/kv/local_test_cluster.go#L50)):*
why do you need this member?

---

*[server/server.go, line 68 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAahi-AIxmS6bFos8MD:-KAahi-AIxmS6bFos8ME:2124961268) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/server/server.go#L68)):*
why do you need this member?

---

*[storage/replica.go, line 785 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAah8U-xYwGxfTLjhhG:-KAah8U-xYwGxfTLjhhH:38851203) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/storage/replica.go#L785)):*
Did you want to pass a context containing this new tracer into the methods below here?

---

*[util/tracing/nettrace.go, line 26 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAahqREhD2TIMkG1sME:-KAahqREhD2TIMkG1sMF:-1507354716) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/util/tracing/nettrace.go#L26)):*
there's some interface being implemented here, right? can you add a static assertion?

---

*[util/tracing/tracer.go, line 47 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAahRcJIgHseoX4nqL1:-KAahRcKVVIQMs-EdMjs:1509317595) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/util/tracing/tracer.go#L47)):*
s/currently //

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385)*
<!-- Sent from Reviewable.io -->