@JackKrupansky 2016-02-15
Will tracing be a user-visible feature (like logging) or strictly a developer tool? I presume the former. If so, it would be good to flag for doc and roadmap.