@nvanbenschoten 2016-02-15
Reviewed 30 of 31 files at r1.
Review status: all files reviewed at latest revision, 10 unresolved discussions, some commit checks failed.

---

*[server/server.go, line 68 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAahi-AIxmS6bFos8MD:-KAakbW2b_VUkLuvSb-t:1984039903) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/server/server.go#L68)):*
It can just be a local var right now, but I'm guessing there are plans to use it in `Server` methods?

---

*[server/server_test.go, line 310 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAaky5ZtRBTbGUKj8Ma:-KAaky5ZtRBTbGUKj8Mb:-1893744714) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/server/server_test.go#L310)):*
What are the benefits/reasons for tracing during tests?

---

*[storage/replica.go, line 785 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAah8U-xYwGxfTLjhhG:-KAalI7HP4U9JoQRGGDm:-1708345501) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/storage/replica.go#L785)):*
I think @tschottdorf mentioned he was going to do most of the context plumbing in a later change.

---

*[util/tracing/nettrace.go, line 26 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAahqREhD2TIMkG1sME:-KAals28TBJX_39cExY6:-729548395) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/util/tracing/nettrace.go#L26)):*
Line 100?

---

*[util/tracing/tracer.go, line 64 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385#-KAam0xWzGCd1luVq830:-KAam0xWzGCd1luVq831:-573117079) ([raw file](https://github.com/cockroachdb/cockroach/blob/df721bd40f4bc7847ed60b565a16aa60fec78359/util/tracing/tracer.go#L64)):*
I find the name `NoopSpan` (as a type or a function name) to be a lot more explicit about the purpose/use. Maybe consider bringing that name back.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4385)*
<!-- Sent from Reviewable.io -->