@tschottdorf 2016-02-15
@petermattis See below. That should amount to the vanilla overhead of `net/trace`.

```
name                   old time/op    new time/op    delta
Insert1_Cockroach-8       471µs ± 3%     460µs ± 1%   -2.50%    (p=0.000 n=9+9)
Insert10_Cockroach-8      834µs ± 3%     881µs ± 9%   +5.62%  (p=0.015 n=10+10)
Insert100_Cockroach-8    4.22ms ± 1%    4.46ms ± 9%   +5.72%   (p=0.002 n=9+10)
Update1_Cockroach-8       981µs ± 1%    1096µs ±26%  +11.76%  (p=0.000 n=10+10)
Update10_Cockroach-8     6.25ms ± 7%    6.29ms ± 4%     ~      (p=0.661 n=10+9)
Update100_Cockroach-8    58.2ms ± 8%    57.2ms ± 1%     ~      (p=0.211 n=10+9)
Delete1_Cockroach-8      1.02ms ± 8%    1.04ms ±11%     ~      (p=0.278 n=10+9)
Delete10_Cockroach-8     1.83ms ± 5%    1.91ms ± 5%   +4.02%  (p=0.043 n=10+10)
Delete100_Cockroach-8    9.13ms ± 1%    9.51ms ±10%   +4.10%  (p=0.029 n=10+10)
Scan1_Cockroach-8         252µs ± 1%     253µs ± 2%     ~       (p=0.321 n=8+9)
Scan10_Cockroach-8        301µs ± 2%     304µs ± 1%   +0.80%   (p=0.010 n=9+10)
Scan100_Cockroach-8       744µs ± 3%     735µs ± 2%     ~      (p=0.065 n=9+10)

name                   old alloc/op   new alloc/op   delta
Insert1_Cockroach-8      25.2kB ± 0%    27.2kB ± 0%   +7.81%  (p=0.000 n=10+10)
Insert10_Cockroach-8     80.2kB ± 0%    82.2kB ± 0%   +2.47%   (p=0.000 n=10+8)
Insert100_Cockroach-8     630kB ± 0%     633kB ± 0%   +0.33%  (p=0.000 n=10+10)
Update1_Cockroach-8      43.0kB ± 0%    49.5kB ± 0%  +15.02%   (p=0.000 n=10+7)
Update10_Cockroach-8      277kB ± 0%     317kB ± 0%  +14.40%  (p=0.000 n=10+10)
Update100_Cockroach-8    2.61MB ± 0%    2.98MB ± 0%  +14.39%  (p=0.000 n=10+10)
Delete1_Cockroach-8      40.5kB ± 0%    47.0kB ± 0%  +15.97%    (p=0.000 n=9+9)
Delete10_Cockroach-8      137kB ± 0%     144kB ± 0%   +4.75%  (p=0.000 n=10+10)
Delete100_Cockroach-8    1.02MB ± 0%    1.02MB ± 0%   +0.62%  (p=0.000 n=10+10)
Scan1_Cockroach-8        11.0kB ± 0%    12.8kB ± 0%  +16.41%  (p=0.000 n=10+10)
Scan10_Cockroach-8       17.3kB ± 0%    19.1kB ± 0%  +10.54%   (p=0.000 n=10+9)
Scan100_Cockroach-8      68.9kB ± 0%    70.7kB ± 0%   +2.64%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
Insert1_Cockroach-8         370 ± 0%       386 ± 0%   +4.32%  (p=0.000 n=10+10)
Insert10_Cockroach-8        881 ± 0%       897 ± 0%   +1.82%   (p=0.000 n=10+8)
Insert100_Cockroach-8     5.73k ± 0%     5.74k ± 0%   +0.30%    (p=0.000 n=9+9)
Update1_Cockroach-8         804 ± 0%       852 ± 0%   +5.97%  (p=0.000 n=10+10)
Update10_Cockroach-8      5.41k ± 0%     5.71k ± 0%   +5.56%   (p=0.000 n=10+8)
Update100_Cockroach-8     51.3k ± 0%     54.1k ± 0%   +5.51%  (p=0.000 n=10+10)
Delete1_Cockroach-8         748 ± 0%       796 ± 0%   +6.46%   (p=0.000 n=8+10)
Delete10_Cockroach-8      1.65k ± 0%     1.70k ± 0%   +2.90%    (p=0.000 n=9+9)
Delete100_Cockroach-8     10.0k ± 0%     10.1k ± 0%   +0.48%  (p=0.000 n=10+10)
Scan1_Cockroach-8           221 ± 0%       234 ± 0%   +5.88%    (p=0.000 n=8+9)
Scan10_Cockroach-8          324 ± 0%       337 ± 0%   +4.08%  (p=0.000 n=10+10)
Scan100_Cockroach-8       1.23k ± 0%     1.25k ± 0%   +1.08%   (p=0.000 n=10+8)
```

I can take out the tracing for these benchmarks for now - after all, it's not the actual tracer yet.